bot_service/requirements.txt
aiogram==3.13.1
aio_pika==9.5.4
python-dotenv~=1.0.1
pillow==11.1.0
aiohttp~=3.10.11
isort==5.13.2
flake8==7.1.1